modules: adapt to new mmap locking API
In 5.8-rc1, mmap locking was reworked to use wrappers around mmap_sem
(which was also renamed to mmap_lock). All code is now supposed to use
these wrappers as the internal implementation of the lock may change in
the future.

Add the wrappers to older kernels to keep the conditionals out of regular
code.