vmmon: work around CentOS 8 build failure
The kernel in CentOS 8 is 4.18 but it seems to contain a backport of
mainline commit ca79b0c211af ("mm: convert totalram_pages and
totalhigh_pages variables to atomic") which switched totalram_pages from
a global variable to a function. As this change is only detected by kernel
version check, we need to add an exception for RHEL / CentOS builds.

Unfortunately none of the CentOS users confirmed which versions this commit
was backported to so I had to resort to a guess that it's 8.0 and newer.
If this turns out to be false, please report.