vmmon: drop pointless check
Compilation with gcc12 warns that check of header->entries for null is
always false because it's an array embedded in struct rather than
a pointer. The check is indeed useless (and omitted by optimizer anyway).
It does no harm either but keeping the build clean of warnings helps to
spot new ones so drop the check.