misc: discontinue Player related versions and tags
For quite long, VMware Workstation and Player have been distributed with
identical host module source. Therefore it seems unnecessary to maintain
two sets of branches and tags and 17.0.0 release is a nice opportunity to
discontinue "player-*" branches and "p*" tags. To reduce the mess, those
will be preserved and created for pre-17 products but no new Player related
branches and tags will be created for versions 17.x and newer unless the
module sources diverge in the future.

Update the README and INSTALL files to reflect this change and replace
14.1.1 version with 17.0.0 in examples.