modules: use explicit module_init() and module_exit() (#187)
Both modules still rely on traditional magic names init_module() and
cleanup_module() for their init and exit functions rather than using
explicit module_init() and module_exit() to mark them Apparently this has
an unfortunate side effect that the two functions are not identified as
indirect call targets by objdump and they get "sealed" when the module is
built against and loaded into an IBT enabled kernel.

Even if this rather looks like an omission on upstream side, rename the
functions and mark them explicitly with module_init() and module_exit() as
it seems considered the preferred style anyway. (The CamelCase names in
vmmon are ugly but consistent with Workstation 17.0.0 code.)