misc: automate DKMS setup and source code updates
DKMS is a convenient framework for managing out-of-tree kernel modules,
such as the ones in this repo. The `dkms-setup.sh` script will add the
modules here (vmmon & vmnet, as of this writing) to the system's DKMS
configuration, allowing them to be automatically rebuilt and resigned
whenever the kernel is upgraded.

CHANGES:
	- Add `dkms.conf` to `.gitignore`. The setup script will
	  generate a `dkms.conf` file from the template `dkms.conf.in`.
	- Add an optional override to the `VM_UNAME` variable in the
	  `Makefile`. This is required to build for any non-running
	  kernel (which DKMS will do whenever a new kernel is installed
	  by the system package manager).

ADDITIONS:
	- `dkms-setup.sh`: Shell script to manage setup and updates of
	  the modules in this repo.
	- `dkms.conf.in`: Template DKMS configuration file to be
	  populated by the setup script.