vmmon: second argument of eventfd_signal() was dropped
Mainline commit 3652117f8548 ("eventfd: simplify eventfd_signal()") drops
second argument of eventfd_signal() as all in-tree callers passed 1.
Fortunately so do we so that we can simply drop it. Unfortunately there
does not seem to be a way to detect this change reliably so that we have to
resort to a version check, risking trouble with backports.